requirements.txt
Django==4.2
djangorestframework==3.14
python-decouple==3.8
python-dateutil==2.8.2
